refactor: split imagetools from module to package
