+++ lektor/imagetools/__init__.py
@@ -1,0 +1,13 @@
+from .exif import read_exif
+from .image_info import get_image_info
+from .thumbnail import make_image_thumbnail
+from .thumbnail import Thumbnail
+from .thumbnail import ThumbnailMode
+
+__all__ = [
+    "get_image_info",
+    "make_image_thumbnail",
+    "read_exif",
+    "Thumbnail",
+    "ThumbnailMode",
+]
