+++ lektor/imagetools/_compat.py
@@ -1,0 +1,34 @@
+"""Compatibility with various versions of Pillow."""
+from __future__ import annotations
+
+from enum import IntEnum
+from types import ModuleType
+from types import SimpleNamespace
+from typing import Mapping
+
+import PIL.ExifTags
+
+__all__ = ["ExifTags", "UnidentifiedImageError"]
+
+PILLOW_VERSION_INFO = tuple(map(int, PIL.__version__.split(".")))
+
+if PILLOW_VERSION_INFO >= (9, 4):
+    ExifTags: ModuleType | SimpleNamespace = PIL.ExifTags
+else:
+    # Pillow < 9.4 does not provide the PIL.ExifTags.{Base,GPS,IFD} enums. Here we provide
+    # and ExifTags namespace which has them.
+
+    def _reverse_map(mapping: Mapping[int, str]) -> dict[str, int]:
+        return dict(map(reversed, mapping.items()))  # type: ignore[arg-type]
+
+    ExifTags = SimpleNamespace(
+        Base=IntEnum("Base", _reverse_map(PIL.ExifTags.TAGS)),
+        GPS=IntEnum("GPS", _reverse_map(PIL.ExifTags.GPSTAGS)),
+        IFD=IntEnum("IFD", [("Exif", 34665), ("GPSInfo", 34853)]),
+        TAGS=PIL.ExifTags.TAGS,
+        GPSTAGS=PIL.ExifTags.GPSTAGS,
+    )
+
+
+# UnidentifiedImageError only exists in Pillow >= 7.0.0
+UnidentifiedImageError = getattr(PIL, "UnidentifiedImageError", OSError)
