+++ tests/imagetools/exif-test-1.jpg
