+++ tests/imagetools/exif-test-2.gif
