+++ tests/imagetools/exif-test-3.jpg
