+++ tests/imagetools/exif-test-4.jpg
