+++ tests/imagetools/test_thumbnail.py
@@ -1,353 +1,36 @@
 from __future__ import annotations
 
 import dataclasses
-import datetime
 import html.parser
 import io
 from contextlib import contextmanager
-from fractions import Fraction
 from pathlib import Path
 from unittest import mock
-from xml.sax import saxutils
 
 import PIL
 import pytest
 from pytest import approx
 
-from lektor.imagetools import _combine_make
-from lektor.imagetools import _compute_cropbox
-from lektor.imagetools import _convert_icc_profile_to_srgb
-from lektor.imagetools import _create_artifact
-from lektor.imagetools import _create_thumbnail
-from lektor.imagetools import _get_thumbnail_url_path
-from lektor.imagetools import _parse_svg_units_px
-from lektor.imagetools import _save_position
-from lektor.imagetools import _to_altitude
-from lektor.imagetools import _to_degrees
-from lektor.imagetools import _to_flash_description
-from lektor.imagetools import _to_float
-from lektor.imagetools import _to_focal_length
-from lektor.imagetools import _to_rational
-from lektor.imagetools import _to_string
-from lektor.imagetools import compute_dimensions
-from lektor.imagetools import CropBox
-from lektor.imagetools import get_image_info
-from lektor.imagetools import ImageSize
-from lektor.imagetools import make_image_thumbnail
-from lektor.imagetools import read_exif
-from lektor.imagetools import Thumbnail
-from lektor.imagetools import ThumbnailMode
-from lektor.imagetools import ThumbnailParams
+from lektor.imagetools.thumbnail import _compute_cropbox
+from lektor.imagetools.thumbnail import _convert_icc_profile_to_srgb
+from lektor.imagetools.thumbnail import _create_artifact
+from lektor.imagetools.thumbnail import _create_thumbnail
+from lektor.imagetools.thumbnail import _get_thumbnail_url_path
+from lektor.imagetools.thumbnail import compute_dimensions
+from lektor.imagetools.thumbnail import CropBox
+from lektor.imagetools.thumbnail import get_image_info
+from lektor.imagetools.thumbnail import ImageSize
+from lektor.imagetools.thumbnail import make_image_thumbnail
+from lektor.imagetools.thumbnail import Thumbnail
+from lektor.imagetools.thumbnail import ThumbnailMode
+from lektor.imagetools.thumbnail import ThumbnailParams
 
 
 HERE = Path(__file__).parent
-DEMO_PROJECT = HERE / "demo-project/content"
+DEMO_PROJECT = HERE / "../demo-project/content"
 ICC_PROFILE_TEST_JPG = DEMO_PROJECT / "icc-profile-test/rgb-to-gbr-test.jpg"
 CMYK_ICC_PROFILE_TEST = DEMO_PROJECT / "icc-profile-test/CGATS001Compat-v2-micro.icc"
 NONIMAGE_FILE_PATH = Path(__file__)  # we are not an image
-
-EXIF_ORIENTATION_TAG = 0x0112
-
-
-@pytest.mark.parametrize(
-    "make, model, expected",
-    [
-        ("Frobnotz", "Frobnicator v42", "Frobnotz Frobnicator v42"),
-        ("Frobnotz", None, "Frobnotz"),
-        (None, "Frobnicator v42", "Frobnicator v42"),
-        ("Frobnotz", "Frobnotz Frobnicator v42", "Frobnotz Frobnicator v42"),
-        (None, None, ""),
-    ],
-)
-def test_combine_make(make, model, expected):
-    assert _combine_make(make, model) == expected
-
-
-@pytest.mark.parametrize(
-    "coerce, value, expected",
-    [
-        (_to_string, "a b", "a b"),
-        (_to_string, "a\xc2\xa0b", "a\xa0b"),
-        (_to_string, "a\xa0b", "a\xa0b"),
-        #
-        (_to_rational, 42, 42),
-        (_to_rational, Fraction("22/7"), Fraction("22/7")),
-        (_to_rational, (3, 2), Fraction("3/2")),
-        #
-        (_to_float, 42, 42),
-        (_to_float, 1.5, 1.5),
-        (_to_float, Fraction("22/7"), approx(3.142857)),
-        (_to_float, (7, 4), 1.75),
-        #
-        (_to_flash_description, 0x41, "Flash fired, red-eye reduction mode"),
-        (_to_flash_description, 0x100, "Flash did not fire (256)"),
-        (_to_flash_description, 0x101, "Flash fired (257)"),
-        (_to_flash_description, -1, "Flash fired (-1)"),
-        #
-        (_to_focal_length, Fraction("45/2"), "22.5mm"),
-        (_to_focal_length, (521, 10), "52.1mm"),
-    ],
-)
-def test_coersion(coerce, value, expected):
-    assert coerce(value) == expected
-
-
-@pytest.mark.parametrize(
-    "coords, hemisphere, expected",
-    [
-        ((Fraction(45), Fraction(15), Fraction(30)), "N", approx(45.2583333)),
-        ((Fraction(45), Fraction(61, 2), Fraction(0)), "S", approx(-45.5083333)),
-        ((122, 0, 0), "W", -122),
-        ((Fraction("45/2"), 0, 0), "E", 22.5),
-        (((45, 2), (30, 1), (0, 1)), "N", approx(23)),
-    ],
-)
-def test_to_degrees(coords, hemisphere, expected):
-    assert _to_degrees(coords, hemisphere) == expected
-
-
-@pytest.mark.parametrize(
-    "altitude, altitude_ref, expected",
-    [
-        (Fraction("1234/10"), b"\x00", approx(123.4)),
-        (Fraction("1234/10"), None, approx(123.4)),
-        ((1234, 10), b"\x00", approx(123.4)),
-        (Fraction("123/10"), b"\x01", approx(-12.3)),
-    ],
-)
-def test_to_altitude(altitude, altitude_ref, expected):
-    assert _to_altitude(altitude, altitude_ref) == expected
-
-
-TEST_JPG_EXIF_INFO = {
-    "altitude": approx(779.0293),
-    "aperture": approx(2.275),
-    "artist": None,
-    "camera": "Apple iPhone 6",
-    "camera_make": "Apple",
-    "camera_model": "iPhone 6",
-    "copyright": None,
-    "created_at": datetime.datetime(2015, 12, 6, 11, 37, 38),
-    "exposure_time": "1/33",
-    "f": "ƒ/2.2",
-    "f_num": approx(2.2),
-    "flash_info": "Flash did not fire, compulsory flash mode",
-    "focal_length": "4.2mm",
-    "focal_length_35mm": "29mm",
-    "iso": 160,
-    "latitude": approx(46.633833),
-    "lens": "Apple iPhone 6 back camera 4.15mm f/2.2",
-    "lens_make": "Apple",
-    "lens_model": "iPhone 6 back camera 4.15mm f/2.2",
-    "longitude": approx(13.404833),
-    "location": approx((46.633833, 13.404833)),
-    "shutter_speed": "1/33",
-    "documentname": "testName",
-    "description": "testDescription",
-    "is_rotated": True,
-}
-
-# This is the default ExifINFO for images without EXIF data
-NULL_EXIF_INFO = {
-    "altitude": None,
-    "aperture": None,
-    "artist": None,
-    "camera": "",
-    "camera_make": None,
-    "camera_model": None,
-    "copyright": None,
-    "created_at": None,
-    "description": None,
-    "documentname": None,
-    "exposure_time": None,
-    "f": None,
-    "f_num": None,
-    "flash_info": None,
-    "focal_length": None,
-    "focal_length_35mm": None,
-    "is_rotated": False,
-    "iso": None,
-    "latitude": None,
-    "lens": "",
-    "lens_make": None,
-    "lens_model": None,
-    "location": None,
-    "longitude": None,
-    "shutter_speed": None,
-}
-
-TEST_IMAGE_EXIF_INFO = {
-    DEMO_PROJECT / "test.jpg": TEST_JPG_EXIF_INFO,
-    DEMO_PROJECT / "test-sof-last.jpg": TEST_JPG_EXIF_INFO,
-    DEMO_PROJECT / "test-progressive.jpg": NULL_EXIF_INFO,
-    HERE / "exif-test-2.gif": NULL_EXIF_INFO,
-    (HERE / "exif-test-1.jpg"): {
-        "altitude": None,
-        "aperture": None,
-        "artist": "Geoffrey T. Dairiki",
-        "camera": "NIKON CORPORATION NIKON D7100",
-        "camera_make": "NIKON CORPORATION",
-        "camera_model": "NIKON D7100",
-        "copyright": "2015",
-        "created_at": datetime.datetime(2022, 10, 22, 9, 20, 56),
-        "description": None,
-        "documentname": None,
-        "exposure_time": "1/400",
-        "f": "ƒ/4.5",
-        "f_num": 4.5,
-        "flash_info": "Flash did not fire, compulsory flash mode",
-        "focal_length": "86mm",
-        "focal_length_35mm": "129mm",
-        "is_rotated": False,
-        "iso": 1250,
-        "latitude": None,
-        "lens": "",
-        "lens_make": None,
-        "lens_model": None,
-        "location": None,
-        "longitude": None,
-        "shutter_speed": None,
-    },
-    HERE / "exif-test-3.jpg": {"altitude": approx(-85.9)},  # negative altitude
-    HERE / "exif-test-4.jpg": {"altitude": approx(85.9)},  # This no GPSAltitudeRef tag:
-}
-
-
-@pytest.mark.parametrize(
-    "path, attr, expected",
-    [
-        (path, attr, expected)
-        for path, tags in TEST_IMAGE_EXIF_INFO.items()
-        for attr, expected in tags.items()
-    ],
-)
-def test_read_exif_attr(path, attr, expected):
-    with path.open("rb") as fp:
-        exif = read_exif(fp)
-    assert getattr(exif, attr) == expected
-
-
-def test_read_exif_unrecognized_image():
-    exif_info = read_exif(io.BytesIO(b"unrecognized-image"))
-    assert not exif_info
-
-
-@pytest.mark.parametrize(
-    "dimension, pixels",
-    [
-        ("123", 123),
-        ("50%", None),
-        ("42px", 42),
-        ("42.5 px", 42.5),
-    ],
-)
-def test_parse_svg_units_px(dimension, pixels):
-    assert _parse_svg_units_px(dimension) == pixels
-
-
-class DummyFileBase:
-    def write(self, fp):
-        raise NotImplementedError()
-
-    def open(self):
-        fp = io.BytesIO()
-        self.write(fp)
-        fp.seek(0)
-        return fp
-
-    def path(self, tmp_path: Path) -> Path:
-        path = tmp_path / "dummy"
-        with path.open("wb") as fp:
-            self.write(fp)
-        return path
-
-
-@dataclasses.dataclass
-class DummyImage(DummyFileBase):
-    width: int
-    height: int
-    format: str = "PNG"
-    orientation: int | None = None
-
-    def write(self, fp):
-        image = PIL.Image.new("RGB", (self.width, self.height), "#999")
-        exif = image.getexif()
-        if self.orientation is not None:
-            exif[EXIF_ORIENTATION_TAG] = self.orientation
-        image.save(fp, self.format, exif=exif)
-
-
-@dataclasses.dataclass
-class DummySVGImage(DummyFileBase):
-    width: int | str | None = None
-    height: int | str | None = None
-    xmlns: str | None = "http://www.w3.org/2000/svg"
-    xml_decl: str | None = '<?xml version="1.0" encoding="UTF-8" standalone="no"?>\n'
-
-    def write(self, fp):
-        values = dataclasses.asdict(self)
-        svg_attrs = "".join(
-            f" {key}={saxutils.quoteattr(str(values[key]))}"
-            for key in ("width", "height", "xmlns")
-            if values[key] is not None
-        )
-        svg = (
-            f'{self.xml_decl or ""}'
-            f"<svg{svg_attrs}>"
-            '<circle cx="50" cy="50" r="50" fill="lightgrey"/>'
-            '<circle cx="30" cy="30" r="10" fill="black"/>'
-            '<circle cx="60" cy="30" r="10" fill="black"/>'
-            '<path stroke="black" d="M30 70 l30 0" stroke-width="5"/>'
-            "</svg>"
-        )
-        fp.write(svg.encode("utf-8"))
-
-
-@dataclasses.dataclass
-class DummyFile(DummyFileBase):
-    content: bytes = b""
-
-    def write(self, fp):
-        fp.write(self.content)
-
-
-@pytest.mark.parametrize(
-    "image, expected",
-    [
-        (DummyFile(b"short"), (None, None, None)),
-        (DummyFile(b"Not an image. " * 40), (None, None, None)),
-        (DummyImage(10, 20, "JPEG"), ("jpeg", 10, 20)),
-        (DummyImage(10, 20, "GIF"), ("gif", 10, 20)),
-        (DummyImage(10, 20, "PNG"), ("png", 10, 20)),
-        (DummyImage(10, 20, "JPEG", orientation=5), ("jpeg", 20, 10)),
-        # check that Exif orientation is ignored for PNG images"
-        (DummyImage(10, 20, "PNG", orientation=7), ("png", 10, 20)),
-        (DummyImage(10, 20, "PPM"), (None, None, None)),
-        (DummySVGImage("10px", "20px"), ("svg", 10, 20)),
-        (DummySVGImage("10", "20", xml_decl=None), ("svg", 10, 20)),
-        (DummySVGImage(None, None), ("svg", None, None)),
-        (DummySVGImage("invalid-width", "invalid-height"), ("svg", None, None)),
-        (DummySVGImage("10px", "10px", xmlns=None), (None, None, None)),
-    ],
-)
-def test_get_image_info(image, expected, tmp_path):
-    image_path = tmp_path / "test-image"
-    with image_path.open("wb") as fp:
-        image.write(fp)
-    assert get_image_info(image_path) == expected
-
-
-def test_get_image_info_fp_deprecated():
-    image = DummyImage(10, 20, "JPEG")
-    with pytest.deprecated_call():
-        assert get_image_info(image.open()) == ("jpeg", 10, 20)
-
-
-def test_save_position():
-    fp = io.BytesIO(b"data")
-    with _save_position(fp) as infp:
-        assert infp.read() == b"data"
-    assert fp.read() == b"data"
 
 
 def test_ThumbnailParams_unrecognized_format():
@@ -668,25 +351,33 @@ def test_make_image_thumbnail_unknown_image_format(tmp_path):
     assert len(ctx.mock_calls) == 0
 
 
-def test_make_image_thumbnail_svg(tmp_path):
+@pytest.fixture
+def dummy_svg_file(tmp_path):
     dummy_svg = tmp_path / "dummy.svg"
-    with dummy_svg.open("wb") as fp:
-        DummySVGImage(width="400px", height="300px").write(fp)
+
+    def svg_file(**kwargs):
+        attrs = " ".join(f'{key}="{value}"' for key, value in kwargs.items())
+        dummy_svg.write_text(f'<svg {attrs} xmlns="http://www.w3.org/2000/svg"></svg>')
+        return dummy_svg
+
+    return svg_file
+
+
+def test_make_image_thumbnail_svg(dummy_svg_file):
+    svg_file = dummy_svg_file(width="400px", height="300px")
     ctx = mock.Mock(name="ctx")
     thumbnail = make_image_thumbnail(
-        ctx, dummy_svg, "/urlpath/dummy.svg", width=80, height=60
+        ctx, svg_file, "/urlpath/dummy.svg", width=80, height=60
     )
     assert (thumbnail.width, thumbnail.height) == (80, 60)
     assert thumbnail.url_path == "/urlpath/dummy.svg"
 
 
-def test_make_image_thumbnail_svg_fit_mode_fails_if_missing_dim(tmp_path):
-    dummy_svg = tmp_path / "dummy.svg"
-    with dummy_svg.open("wb") as fp:
-        DummySVGImage(width="400px", height=None).write(fp)
+def test_make_image_thumbnail_svg_fit_mode_fails_if_missing_dim(dummy_svg_file):
+    svg_file = dummy_svg_file(width="400px")  # missing height
     ctx = mock.Mock(name="ctx")
     with pytest.raises(ValueError, match="Cannot determine aspect ratio"):
-        make_image_thumbnail(ctx, dummy_svg, "/urlpath/dummy.svg", width=80, height=60)
+        make_image_thumbnail(ctx, svg_file, "/urlpath/dummy.svg", width=80, height=60)
 
 
 def test_Thumbnail_str():
@@ -780,31 +471,3 @@ class TestFunctional:
     def test_large_thumbnail_returns_original(self, built_demo_img_tags):
         large_thumb = ImgTag(alt="original", src="test.jpg", width=384, height=512)
         assert large_thumb in built_demo_img_tags
-
-
-@pytest.fixture(
-    params=[
-        "test.jpg",  # base image, exif-rotated
-        "test-sof-last.jpg",  # same image but with SOF marker last
-        "test-progressive.jpg",  # with progressive encoding, rotated in place
-    ]
-)
-def demo_test_jpg(request, pad):
-    return pad.root.attachments.images.get(request.param)
-
-
-def test_image_attributes(demo_test_jpg):
-    assert demo_test_jpg.width == 384
-    assert demo_test_jpg.height == 512
-    assert demo_test_jpg.format == "jpeg"
-
-
-def test_exif(demo_test_jpg):
-    name = demo_test_jpg.path.lstrip("/")
-    expected = TEST_IMAGE_EXIF_INFO[DEMO_PROJECT / name]
-
-    assert demo_test_jpg.exif.to_dict() == expected
-    if expected == NULL_EXIF_INFO:
-        assert not demo_test_jpg.exif
-    else:
-        assert demo_test_jpg.exif
